refactor install module
